+++ RELEASES.md
@@ -1,6 +1,7 @@
 # Version 0.1.6 (2025-11-05)
 
   * Fix KaTeX, maybe?
+  * Rename `Graph` to `Tree`.
   * Fix text form for `"{:-}"`.
 
 # Version 0.1.5 (2025-11-04)